+++ crates/apollo-compiler/src/validation/mod.rs
@@ -293,6 +293,7 @@ impl DiagnosticData {
                     QueryRootOperationType => "QueryRootOperationType",
                     UnusedVariable { .. } => "UnusedVariable",
                     RootOperationObjectType { .. } => "RootOperationObjectType",
+                    DuplicateRootOperationType { .. } => "DuplicateRootOperationType",
                     UnionMemberObjectType { .. } => "UnionMemberObjectType",
                     UnsupportedLocation { .. } => "UnsupportedLocation",
                     UnsupportedValueType { .. } => "UnsupportedValueType",
@@ -350,6 +351,14 @@ impl DiagnosticData {
                 }
                 ExecutableBuildError::SubscriptionUsesConditionalSelection { .. } => {
                     "SubscriptionUsesConditionalSelection"
+                }
+                ExecutableBuildError::DuplicateDeferLabel { .. } => "DuplicateDeferLabel",
+                ExecutableBuildError::DeferLabelMustNotBeVariable => "DeferLabelMustNotBeVariable",
+                ExecutableBuildError::DeferOnRootMutationOrSubscriptionField { .. } => {
+                    "DeferOnRootMutationOrSubscriptionField"
+                }
+                ExecutableBuildError::DeferInSubscriptionMustBeConditional => {
+                    "DeferInSubscriptionMustBeConditional"
                 }
                 ExecutableBuildError::ConflictingFieldType(_) => "ConflictingFieldType",
                 ExecutableBuildError::ConflictingFieldName(_) => "ConflictingFieldName",
@@ -435,6 +444,7 @@ impl DiagnosticData {
                         Some(format!(r#"Variable "${name}" is never used."#))
                     }
                     RootOperationObjectType { .. } => None,
+                    DuplicateRootOperationType { .. } => None,
                     UnionMemberObjectType { .. } => None,
                     UnsupportedLocation { name, location, .. } => Some(format!(
                         r#"Directive "@{name}" may not be used on {location}."#
@@ -688,6 +698,22 @@ impl DiagnosticData {
                         r#"Fields "{alias}" conflict because they have differing arguments. Use different aliases on the fields to fetch both if this was intentional."#
                     ))
                 }
+                ExecutableBuildError::DuplicateDeferLabel { .. } => Some(
+                    r#"Value for arguments "defer(label:)" and "stream(label:)" must be unique across all Defer/Stream directive usages."#
+                        .to_string(),
+                ),
+                ExecutableBuildError::DeferLabelMustNotBeVariable => Some(
+                    r#"Argument "@defer(label:)" must be a static string."#.to_string(),
+                ),
+                ExecutableBuildError::DeferOnRootMutationOrSubscriptionField {
+                    operation_type,
+                } => Some(format!(
+                    r#"Defer directive cannot be used on root {} type."#,
+                    operation_type.name()
+                )),
+                ExecutableBuildError::DeferInSubscriptionMustBeConditional => Some(
+                    "Defer directive not supported on subscription operations. Disable `@defer` by setting the `if` argument to `false`.".to_string(),
+                ),
             },
             _ => None,
         }
@@ -914,6 +940,27 @@ impl ToCliReport for DiagnosticData {
                 }
                 ExecutableBuildError::SubscriptionUsesConditionalSelection { .. } => {
                     report.with_label_opt(self.location, "conditional directive used here");
+                }
+                ExecutableBuildError::DuplicateDeferLabel {
+                    original_location, ..
+                } => {
+                    report.with_label_opt(*original_location, "label first used here");
+                    report.with_label_opt(self.location, "duplicate label used here");
+                }
+                ExecutableBuildError::DeferLabelMustNotBeVariable => {
+                    report.with_label_opt(self.location, "variable used as label here");
+                }
+                ExecutableBuildError::DeferOnRootMutationOrSubscriptionField { operation_type } => {
+                    report.with_label_opt(
+                        self.location,
+                        format_args!("`@defer` on a root {} selection", operation_type.name()),
+                    );
+                }
+                ExecutableBuildError::DeferInSubscriptionMustBeConditional => {
+                    report.with_label_opt(
+                        self.location,
+                        "`@defer` in a subscription must use `if: false` or `if: $variable`",
+                    );
                 }
                 ExecutableBuildError::ConflictingFieldType(inner) => {
                     let ConflictingFieldType {